+++ pkg/config/config_test.go
@@ -85,6 +85,7 @@ func TestMerge(t *testing.T) {
 		Path:                "some-other",
 		Verbose:             true,
 		Debug:               true,
+		NoColor:             true,
 		IgnoreDefaults:      true,
 		SpacesAftertabs:     true,
 		Exclude:             []string{"some-other"},
@@ -106,9 +107,11 @@ func TestMerge(t *testing.T) {
 	expected := mergeConfig
 	expected.Logger.Verbosee = true
 	expected.Logger.Debugg = true
+	expected.Logger.NoColor = true
 
 	if !reflect.DeepEqual(c1, &expected) {
 		t.Errorf("%#v", &expected)
+		t.Errorf("%#v", c1)
 		t.Errorf("Expected, got %#v and %#v", c1, &expected)
 	}